+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## v0.0.24 [unreleased]
+
+### Features
+
+- Create AMI by importing VM image from S3: `awless import image bucket=my-bucket s3object=my-object`
+- Listing pending import image tasks with `awless list importimagetasks`
+- Deleting images and optionally its related snapshots `awless delete image delete-snapshots=true`
+
 ## v0.0.23 [2017-05-05]
 
 ### Features